Fix tests, reorganise modules
