+++ bin/delete-webhook.js
@@ -1,2 +1,2 @@
 #! /usr/bin/env node
-require('../dist').utils.deleteWebhook();
+require('../dist').deleteWebhook();
