+++ bin/set-webhook.js
@@ -1,6 +1,6 @@
 #! /usr/bin/env node
 const [, , url, maxConnections, ...allowedUpdates] = process.argv;
-require('../dist').utils.setWebhook({
+require('../dist').setWebhook({
   url,
   max_connections: parseInt(maxConnections),
   allowed_updates: allowedUpdates.length && allowedUpdates,
