+++ jest.config.js
@@ -9,4 +9,6 @@ module.exports = {
   collectCoverage: true,
   // make sure to include coverage from all files even those not tested at all
   collectCoverageFrom: ['src/**/*.{ts,js}'],
+  // do setup steps (silenece debug & info console logs)
+  setupFilesAfterEnv: ['<rootDir>/test/setup.ts'],
 };
