+++ src/index.ts
@@ -1,44 +1,18 @@
-import { APIGatewayProxyHandlerV2 } from 'aws-lambda';
-import { AzureHttpFunction, wrapAws, wrapAzure } from './wrap-http';
-import wrapTelegram from './wrap-telegram';
+import wrapTelegram from './telegram/wrap-telegram';
 import type {
+  AwsHttpFunction,
+  AzureHttpFunction,
   HandlerMap,
   MessageHandler,
-  UpdateResponse,
-} from './wrap-telegram/types';
+} from './types';
+import { wrapAws, wrapAzure } from './wrap-http';
 
 // re-exports
 export { DevServer, startDevServer } from './dev-server';
+export { Env, InlineEnv, MessageEnv } from './telegram/env';
+export { deleteWebhook, setWebhook } from './telegram/webhook-utils';
+export * from './types';
 export * as utils from './utils';
-export type {
-  AzureHttpFunction,
-  BodyHandler,
-  Context,
-  HttpRequest,
-  HttpResponse,
-  Logger,
-} from './wrap-http';
-export { Env, InlineEnv, MessageEnv } from './wrap-telegram/env';
-export type {
-  AnswerInlineQuery,
-  Chat,
-  FileBuffer,
-  HandlerMap,
-  InlineHandler,
-  InlineQuery,
-  InlineQueryResult,
-  InlineResult,
-  InputFile,
-  Message,
-  MessageHandler,
-  MessageResponse,
-  NoResponse,
-  ResponseMethod,
-  ResponseObject,
-  TgApiRequest,
-  Update,
-  User,
-} from './wrap-telegram/types';
 export { wrapAzure, wrapAws, wrapTelegram };
 
 // single combined wrapper for convenience
@@ -50,7 +24,6 @@ export const createAzureTelegramWebhook = (
 export const createAwsTelegramWebhook = (
   handler: MessageHandler | HandlerMap,
   errorChatId?: number,
-): APIGatewayProxyHandlerV2<UpdateResponse> =>
-  wrapAws(wrapTelegram(handler, errorChatId));
+): AwsHttpFunction => wrapAws(wrapTelegram(handler, errorChatId));
 
 export default createAzureTelegramWebhook;
