+++ test/__test-project-aws__/template.yml
