+++ test/index.test.ts
@@ -1,20 +1,26 @@
-import { ctx } from './helpers';
-import type {
-  HttpRequest,
+import {
+  AwsHttpRequest,
+  AzureHttpRequest,
+  createAwsTelegramWebhook,
+  createAzureTelegramWebhook,
   Message,
   MessageHandler,
   MessageResponse,
 } from '../src';
-import { createAzureTelegramWebhook } from '../src';
+import { awsCtx, azureCtx } from './helpers';
 
-const textMsgUpdate = {
+const azureRequest = {
   body: {
     update_id: 1,
     message: { text: 'more good things please', chat: { id: 1 } },
   },
-} as HttpRequest;
+} as AzureHttpRequest;
 
-const textResponse = {
+const awsRequest = {
+  body: JSON.stringify(azureRequest.body),
+} as AwsHttpRequest;
+
+const azureResponse = {
   body: {
     chat_id: 1,
     method: 'sendMessage',
@@ -25,11 +31,22 @@ const textResponse = {
   },
 };
 
+const awsResponse = azureResponse.body;
+
 const handler: MessageHandler = ({ text }: Message): MessageResponse => text;
 
 describe('createAzureTelegramWebhook', () => {
   it('handles telegram webhook updates and responses via http', () => {
     const webhook = createAzureTelegramWebhook(handler);
-    return expect(webhook(ctx, textMsgUpdate)).resolves.toEqual(textResponse);
+    return expect(webhook(azureCtx, azureRequest)).resolves.toEqual(
+      azureResponse,
+    );
+  });
+});
+
+describe('createAwsTelegramWebhook', () => {
+  it('handles telegram webhook updates and responses via http', () => {
+    const webhook = createAwsTelegramWebhook(handler);
+    return expect(webhook(awsRequest, awsCtx)).resolves.toEqual(awsResponse);
   });
 });
