+++ test/wrap-http.test.ts
@@ -1,13 +1,12 @@
-import { HttpRequest } from '@azure/functions';
+import type { AzureHttpRequest, AzureHttpResponse } from '../src';
 import { wrapAws, wrapAzure } from '../src';
-import { HttpResponse } from '../src/wrap-http';
-import { ctx } from './helpers';
+import { awsCtx, azureCtx } from './helpers';
 
 const jsonObj = {
   message: { text: 'more good things please', chat: { id: 1 } },
 };
 
-const jsonRequest: HttpRequest = {
+const jsonRequest: AzureHttpRequest = {
   method: 'POST',
   url: 'http://localhost:7071/api/telegram-webhook',
   headers: {
@@ -23,7 +22,7 @@ const jsonRequest: HttpRequest = {
   rawBody: '{"message":{"text":"more good things please","chat":{"id":1}}}',
 };
 
-const stringRequest: HttpRequest = {
+const stringRequest: AzureHttpRequest = {
   method: 'POST',
   url: 'http://localhost:7071/api/telegram-webhook',
   headers: {
@@ -38,7 +37,7 @@ const stringRequest: HttpRequest = {
   rawBody: 'text body',
 };
 
-const emptyRequest: HttpRequest = {
+const emptyRequest: AzureHttpRequest = {
   method: 'POST',
   url: 'http://localhost:7071/api/telegram-webhook',
   headers: {
@@ -57,32 +56,32 @@ describe('wrapAzure', () => {
   const echo = wrapAzure(async (x: any) => x);
 
   it('handles JSON objects', () => {
-    return expect(echo(ctx, jsonRequest)).resolves.toEqual({
+    return expect(echo(azureCtx, jsonRequest)).resolves.toEqual({
       body: jsonObj,
       headers: { 'Content-Type': 'application/json' },
     });
   });
 
   it('works with plain strings', () => {
-    return expect(echo(ctx, stringRequest)).resolves.toEqual({
+    return expect(echo(azureCtx, stringRequest)).resolves.toEqual({
       body: 'text body',
     });
   });
 
   it('ignores empty body', () => {
-    return expect(echo(ctx, emptyRequest)).resolves.toBeUndefined();
+    return expect(echo(azureCtx, emptyRequest)).resolves.toBeUndefined();
   });
 
   it('passes through user defined HTTP responses', () => {
-    let res: HttpResponse = { status: 204, headers: { bar: 'baz' } };
+    let res: AzureHttpResponse = { status: 204, headers: { bar: 'baz' } };
     // deep copy in case the response object is modified
     const returnsRes = wrapAzure(async () => JSON.parse(JSON.stringify(res)));
-    return expect(returnsRes(ctx, jsonRequest)).resolves.toEqual(res);
+    return expect(returnsRes(azureCtx, jsonRequest)).resolves.toEqual(res);
   });
 
   it('sets ctx.res as well as the return value', async () => {
-    const returnVal = await echo(ctx, stringRequest);
-    expect(ctx.res).toEqual(returnVal);
+    const returnVal = await echo(azureCtx, stringRequest);
+    expect(azureCtx.res).toEqual(returnVal);
   });
 });
 
@@ -91,11 +90,11 @@ describe('wrapAws', () => {
 
   it('parses Json', () => {
     return expect(
-      echo({ body: JSON.stringify(jsonObj) } as any, null, null),
+      echo({ body: JSON.stringify(jsonObj) } as any, awsCtx),
     ).resolves.toEqual(jsonObj);
   });
 
   it('ignores empty body and returns empty string instead of undefined', () => {
-    return expect(echo({} as any, null, null)).resolves.toEqual('');
+    return expect(echo({} as any, awsCtx)).resolves.toEqual('');
   });
 });
